move easing files to detail namespace
