+++ include/tweeny/detail/easing/linear.h
@@ -22,58 +22,33 @@ IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-#ifndef TWEENY_EASING_DEF_H
-#define TWEENY_EASING_DEF_H
+#ifndef TWEENY_DETAIL_EASING_LINEAR_H
+#define TWEENY_DETAIL_EASING_LINEAR_H
 
 #include <type_traits>
 #include <cmath>
-#include <utility>
 
 namespace tweeny::detail {
-  template <class...> struct voidify {
-    using type = void;
-  };
-
-  template <class... Ts> using void_t = typename voidify<Ts...>::type;
-
-  template <class, class = void>
-  struct supports_arithmetic_operations : std::false_type {};
-
-  template <class T>
-  struct supports_arithmetic_operations<T, void_t<
-                                          decltype(std::declval<T>() + std::declval<T>()),
-                                          decltype(std::declval<T>() - std::declval<T>()),
-                                          decltype(std::declval<T>() * std::declval<T>()),
-                                          decltype(std::declval<T>() * std::declval<float>()),
-                                          decltype(std::declval<float>() * std::declval<T>())
-                                        >> : std::true_type {};
-
-  struct defaultEasing {
+  struct linearEasing {
     template <typename T>
-    static std::enable_if_t<std::is_integral_v<T>, T> run(float position, T start, T end) {
+    static std::enable_if_t<std::is_integral_v<T>, T> run(const float position, T start, T end) {
       return static_cast<T>(roundf((end - start) * position + start));
     }
 
     template <typename T>
-    static typename std::enable_if<supports_arithmetic_operations<T>::value && !std::is_integral<T>::value, T>::type
-      run(float position, T start, T end) {
+    static std::enable_if_t<!std::is_integral_v<T>, T> run(const float position, T start, T end) {
       return static_cast<T>((end - start) * position + start);
     }
 
     template <typename T>
-    static std::enable_if_t<!supports_arithmetic_operations<T>::value, T> run(float /*position*/, T start, T /*end*/) {
-      return start;
-    }
-
-    template <typename T>
-    T operator()(float position, T start, T end) const {
+    T operator()(const float position, T start, T end) const {
       return run<T>(position, start, end);
     }
   };
 }
 
 namespace tweeny::easing {
-  inline constexpr detail::defaultEasing def{};
+  inline constexpr detail::linearEasing linear{};
 }
 
-#endif // TWEENY_EASING_DEF_H
+#endif // TWEENY_DETAIL_EASING_LINEAR_H
