+++ include/tweeny/detail/easing/stepped.h
@@ -22,41 +22,25 @@ IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-#ifndef TWEENY_INTERPOLATE_H
-#define TWEENY_INTERPOLATE_H
-#include <cstddef>
-#include <tuple>
-#include <utility>
-
-#include "key-frame.h"
-#include "easing.h"
+#ifndef TWEENY_DETAIL_EASING_STEPPED_H
+#define TWEENY_DETAIL_EASING_STEPPED_H
 
 namespace tweeny::detail {
-  template <std::size_t I, typename... ValueTypes>
-  static auto interpolate_one(
-    float t,
-    const key_frame<ValueTypes...> & base,
-    const key_frame<ValueTypes...> & next
-  ) -> std::remove_reference_t<decltype(std::get<I>(base.values))> {
-    const auto & start = std::get<I>(base.values);
-    const auto & end = std::get<I>(next.values);
-    const auto & func = std::get<I>(base.easing_functions);
-    if (func) return func(t, start, end);
-    return easing::def(t, start, end);
-  }
+  struct steppedEasing {
+    template <typename T>
+    static T run(float /*position*/, T start, T /*end*/) {
+      return start;
+    }
 
-  template <typename... ValueTypes, std::size_t... I>
-  static auto interpolate_values(
-    float t,
-    const key_frame<ValueTypes...> & base,
-    const key_frame<ValueTypes...> & next,
-    std::index_sequence<I...>
-  ) -> typename key_frame<ValueTypes...>::values_t {
-    using values_t = typename key_frame<ValueTypes...>::values_t;
-    values_t out{};
-    ((std::get<I>(out) = interpolate_one<I>(t, base, next)), ...);
-    return out;
-  }
+    template <typename T>
+    T operator()(const float position, T start, T end) const {
+      return run<T>(position, start, end);
+    }
+  };
 }
 
-#endif //TWEENY_INTERPOLATE_H
+namespace tweeny::easing {
+  inline constexpr detail::steppedEasing stepped{};
+}
+
+#endif // TWEENY_DETAIL_EASING_STEPPED_H
