+++ include/tweeny/detail/key-frame.h
@@ -22,8 +22,8 @@ IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-#ifndef TWEENY_KEY_FRAME_H
-#define TWEENY_KEY_FRAME_H
+#ifndef TWEENY_DETAIL_KEY_FRAME_H
+#define TWEENY_DETAIL_KEY_FRAME_H
 
 
 #include <algorithm>
@@ -128,4 +128,4 @@ namespace tweeny::detail {
   };
 }
 
-#endif //TWEENY_KEY_FRAME_H
+#endif //TWEENY_DETAIL_KEY_FRAME_H
