+++ include/tweeny/detail/tuple-utilities.h
@@ -22,8 +22,8 @@ IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-#ifndef TWEENY_TUPLE_UTILITIES_H
-#define TWEENY_TUPLE_UTILITIES_H
+#ifndef TWEENY_DETAIL_TUPLE_UTILITIES_H
+#define TWEENY_DETAIL_TUPLE_UTILITIES_H
 #include <cstddef>
 #include <tuple>
 
@@ -58,4 +58,4 @@ namespace tweeny::detail {
   }
 }
 
-#endif //TWEENY_TUPLE_UTILITIES_H
+#endif //TWEENY_DETAIL_TUPLE_UTILITIES_H
