+++ include/tweeny/detail/tween-value.h
@@ -22,8 +22,8 @@ IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-#ifndef TWEENY_TWEEN_VALUE_H
-#define TWEENY_TWEEN_VALUE_H
+#ifndef TWEENY_DETAIL_TWEEN_VALUE_H
+#define TWEENY_DETAIL_TWEEN_VALUE_H
 
 #include <type_traits>
 #include "value-container.h"
@@ -62,4 +62,4 @@ namespace tweeny::detail {
 
 }
 
-#endif // TWEENY_TWEEN_VALUE_H
+#endif // TWEENY_DETAIL_TWEEN_VALUE_H
