+++ include/tweeny/detail/value-container.h
@@ -22,8 +22,8 @@ IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-#ifndef TWEENY_VALUE_CONTAINER_H
-#define TWEENY_VALUE_CONTAINER_H
+#ifndef TWEENY_DETAIL_VALUE_CONTAINER_H
+#define TWEENY_DETAIL_VALUE_CONTAINER_H
 
 #include <array>
 #include <type_traits>
@@ -92,4 +92,4 @@ namespace tweeny::detail {
     >;
 }
 
-#endif //TWEENY_VALUE_CONTAINER_H
+#endif //TWEENY_DETAIL_VALUE_CONTAINER_H
